+++ python_src/leb/freeze/calibration.py
@@ -1,4 +1,31 @@
-"""Calibration routines for Fourier Ptychography"""
+"""Calibration routines for Fourier Ptychography.
+
+Calibration requires a clear understanding of the coordinate systems that are involved in LED
+matrix microscopy. One important coordinate system is the local coordinate system of the matrix.
+Another important one is the global coordinate system of the instrument.
+
+The global coordinate system has its origin where the sample plane intersects the optics axis
+(assuming the sample is perpendicular to the optics axis). The positive z-direction points 
+towards the microscope, and the negative z-direction points towards the LED matrix. The matrix is
+located at z=axial_offset, which is negative.
+
+The local coordinate system of the matrix has its origin at the upper left LED, looking at the
+LED-side of the matrix. The positive x-direction points to the right, and the positive y-direction
+points down. This is the standard in computer graphics. However, this means that, if the coordinate
+system is right-handed, then the positive z-direction points into the matrix. This is the opposite
+of the global coordinate system.
+
+To transform from the local coordinate system of the matrix to the global coordinate system, we
+first rotate an LED coordinate in the local coordinate system to align with the global x-axis.
+Then, we negate the y-axis of the local coordinate system to align with the global y-axis. This is
+the same as rotating by 180 degrees about the intermediate x-axis, which flips both the y- and
+z-axes.
+
+Positive rotations are in the counterclockwise direction about the positive z-axis of a coordinate
+system. Rotations are active, meaning that a point is rotated from one coordinate system to
+another, rather than the coordinate system being rotated to align with another.
+
+"""
 
 import numpy as np
 from numpy.testing import assert_array_almost_equal_nulp
@@ -21,16 +48,12 @@ def calibrate_rectangular_matrix(
     axial_offset_mm: float = -50e3,
     rot_deg: float = 0.0,
     wavelength_um: float = 0.488,
+    flip_yz: bool = True,
     sort: bool = False,
 ) -> Calibration:
     """Computes the wavevectors that correspond to a set of LED coordinates on a rectangular matrix.
 
-    This function requires two coordinate systems. The first is the local coordinate system of the
-    matrix, and the second is the global coordinate system of the instrument.
-
-    The global coordinate system has its origin at the sample z=0 plane. The positive z-direction
-    points towards the microscope, and the negative z-direction points towards the LED matrix. The
-    matrix is located at z=axial_offset.
+    See the module docstring for a description of the coordinate systems.
 
     The wavevectors are computed after the coordinate system of LED indexes is converted to the
     global coordinate system.
@@ -38,21 +61,28 @@ def calibrate_rectangular_matrix(
     Parameters
     ----------
     led_indexes : list[LEDIndexes]
-        The (col, row) indexes of the LEDs in the matrix.
+        The (x, y) indexes of the LEDs in the matrix.
     center_led : LEDIndexes
-        The (col, row) indexes of the center LED.
+        The (x, y) indexes of the center LED. The center LED is determined from a calibration
+        procedure in which it is brought on to the optics axis.
     pitch_mm : float | tuple[float, float]
         The horizontal/vertical distance between LEDs. If a single number, then the pitch is
         assumed uniform in the horizontal and vertical directions. If a tuple, then the first
         number is the horizontal (x) distance and the second is the vertical (y) distance.
     lateral_offset_mm : tuple[float, float]
-        The (x, y) offset from the origin of the global coordinate system to the origin of the
-        matrix coordinate system.
+        The (x, y) offset from the origin of the global coordinate system to the center LED
+        coordinates of the matrix. This can be used to account for the fact that the center LED
+        may not be perfectly centered on the optics axis after calibration.
     axial_offset_mm : float
         The offest from the LED matrix to the sample, which lies at z = 0.
-    rot_deg : float | tuple[float, float]
-        The rotation of the matrix about its central z-axis in degrees. +z points away from the
-        surface with the LEDs.
+    rot_deg : float
+        The rotation of the matrix about its central z-axis in degrees. Positive rotations are
+        clockwise when looking at the LED side of the matrix because the z-axis of the local
+        coordinate system points into the matrix.
+    flip_yz : bool
+        If True, then the y- and z-axes are rotated by 180 degrees about the x-axis. This is
+        necessary if the local coordinate system of the matrix is right-handed and the +z-axis
+        points into the matrix, which is the standard in computer graphics
     wavelength_um : float
         The center wavelength of light emitted from the LEDs.
     sort : bool
@@ -71,33 +101,41 @@ def calibrate_rectangular_matrix(
     if not isinstance(pitch_mm, tuple):
         pitch_mm = (pitch_mm, pitch_mm)
 
-    # Compute the LED x, y coordinates from the LED indexes
+    # Translate the origin of the LED matrix coordinate system to the center LED
     led_coords = np.array(led_indexes)
     led_coords[:, 0] = (led_coords[:, 0] - center_led[0]) * pitch_mm[0]
     led_coords[:, 1] = (led_coords[:, 1] - center_led[1]) * pitch_mm[1]
 
-    # Transform the LED coordinates to the global coordinate system
+    # Rotate the LED coordinates about the z-axis to align the local x-axis with the global x-axis.
+    # The rotation angle is negated because we specify the rotation of the matrix relative to its
+    # position with the x-axis aligned with the global x-axis, but the rotation operation is
+    # performed in the opposite direction.
+    rot_deg = -rot_deg
     rotation_matrix = np.array(
         [
             [np.cos(np.deg2rad(rot_deg)), -np.sin(np.deg2rad(rot_deg))],
             [np.sin(np.deg2rad(rot_deg)), np.cos(np.deg2rad(rot_deg))],
         ]
     )
-    led_coords = np.matmul(led_coords, rotation_matrix) + np.array(lateral_offset_mm)
+    intermediate_coords = np.matmul(led_coords, rotation_matrix)
 
-    # Convert to microns
-    led_coords /= 1e3
-    axial_offset_um = axial_offset_mm / 1e3
+    # Flip the y-axis to align the local y-axis with the global y-axis and account for any lateral
+    # offset
+    led_coords = intermediate_coords
+    if flip_yz:
+        led_coords[:, 1] = -led_coords[:, 1]
+    led_coords += np.array(lateral_offset_mm)
 
-    # Compute the wavevectors; direction cosines are the negative because the matrix is behind the
+    # Compute the wavevectors; direction cosines are negative because the matrix is behind the
     # sample.
     k = 2 * np.pi / wavelength_um
-    R = np.sqrt(led_coords[:, 0] ** 2 + led_coords[:, 1] ** 2 + axial_offset_um**2)
-    dir_cos_x = -led_coords[:, 0] / R
+    R = np.sqrt(led_coords[:, 0] ** 2 + led_coords[:, 1] ** 2 + axial_offset_mm**2)
+    dir_cos_x = -led_coords[:, 0] / R  # dimensionless b/c led_coords and R are in mm
     dir_cos_y = -led_coords[:, 1] / R
+    dir_cos_z = -axial_offset_mm / R
     kx = k * dir_cos_x
     ky = k * dir_cos_y
-    kz = -k * axial_offset_um / R
+    kz = k * dir_cos_z
 
     # TODO Modify direction cosines to account for the refraction in the glass.